Moved Pyramid to a completely different strategy, see description
The library started out as having a completely separate component for every block. Creating new components would require the user from duplicating the component and modifying it to their requirements. This works, but I’ve noticed that a lot of components use the same fields.

Say you create a ‘Input and content’ component, this component has two fields, one <input> field and a content field (<textarea>). Another component say ‘Image and content’ would use the same content field. By making them completely different components they would create duplicate code (which could cause errors in one component but not the other) and different components do need tests for each one of them.

The new strategy is to use a schema object, this object defines which fields are shown in the components. The fields are shared between components. This means we only have to test one field and that we’re able to easily create new components without changing code. The only downside is that a component is not as easily configurable without affecting other components.

In this commit you could see that the components in the components folder are removed and placed under the Components/Fields folder. Instead of multiple components there’s now a single one (Customisable) that digests the schema objects defined in the blockList.js file.